refactoring networking logic for updating players
we have a richer packet which contains position as well as the state of
the entity / player.

also renamed various members to make them conform more with that which
they are doing now.

I believe there was a UB bug with args->all_players_connection_info
which casued the local runtime of the server to change the first infex
of the array and break it.

Moved it out of the struct and scoped it alone inside of run_server as
it really shouldn't be a member of the struct anyhow.